Remove partly IntegerRing. TO COMPLETE Add a test over MBAInversion: needs to add proper Matrix Equality is modp fields

+++ src/test/java/bla/linearalgebra/polynomial/TestIntegerPolynomialFFTMulRing.java
@@ -1,5 +1,7 @@
 package bla.linearalgebra.polynomial;
 
+import java.math.BigInteger;
+
 import bla.linearalgebra.IRing;
 import bla.linearalgebra.polynomial.impl.PolynomialAdditionOperation;
 import bla.linearalgebra.polynomial.impl.PolynomialMultiplication_FFT;
@@ -8,8 +10,8 @@ import bla.linearalgebra.polynomial.impl.PolynomialRing;
 public class TestIntegerPolynomialFFTMulRing extends TestIntegerPolynomialModPRing {
 
 	@Override
-	public IRing<IPolynomialFunction<Integer>> makeRing() {
-		return new PolynomialRing<Integer>(new PolynomialAdditionOperation<Integer>(makeCoeffRing()),
-				new PolynomialMultiplication_FFT<Integer>(makeCoeffRing()));
+	public IRing<IPolynomialFunction<BigInteger>> makeRing() {
+		return new PolynomialRing<BigInteger>(new PolynomialAdditionOperation<BigInteger>(makeCoeffRing()), new PolynomialMultiplication_FFT<BigInteger>(
+				makeCoeffRing()));
 	}
 }
