+++ src/test/java/bla/linearalgebra/polynomial/TestZeroMonomial.java
@@ -3,6 +3,8 @@ package bla.linearalgebra.polynomial;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 
+import java.math.BigInteger;
+
 import org.junit.Test;
 
 import bla.linearalgebra.polynomial.monomial.impl.ZeroMonomialFunction;
@@ -11,7 +13,7 @@ public class TestZeroMonomial {
 	protected TestIntegerPolynomialRing parentTestClass = new TestIntegerPolynomialRing();
 
 	public IPolynomialFunction<?> makeZero() {
-		return new ZeroMonomialFunction<Integer>(parentTestClass.makeCoeffRing());
+		return new ZeroMonomialFunction<BigInteger>(parentTestClass.makeCoeffRing());
 	}
 
 	@Test
