+++ src/test/java/bla/linearalgebra/polynomial/string/TestIntegerPolynomialToString.java
@@ -1,5 +1,7 @@
 package bla.linearalgebra.polynomial.string;
 
+import java.math.BigInteger;
+
 import junit.framework.Assert;
 
 import org.junit.Test;
@@ -14,8 +16,8 @@ public class TestIntegerPolynomialToString {
 
 	@Test
 	public void testToString() {
-		IPolynomialFunction<Integer> threePowerTwo = new MonomialFunction<Integer>(parentTestClass.makeCoeffRing(), 3, 2);
-		IPolynomialFunction<Integer> fivePowerThree = new MonomialFunction<Integer>(parentTestClass.makeCoeffRing(), 5, 3);
+		IPolynomialFunction<BigInteger> threePowerTwo = new MonomialFunction<BigInteger>(parentTestClass.makeCoeffRing(), 3, 2);
+		IPolynomialFunction<BigInteger> fivePowerThree = new MonomialFunction<BigInteger>(parentTestClass.makeCoeffRing(), 5, 3);
 
 		Assert.assertEquals("3.2", threePowerTwo.toString());
 		Assert.assertEquals("3.2+5.3", parentTestClass.makeRing().add(threePowerTwo, fivePowerThree).toString());
@@ -24,8 +26,8 @@ public class TestIntegerPolynomialToString {
 
 	@Test
 	public void testFromString() {
-		IPolynomialFunction<Integer> threePowerTwo = new MonomialFunction<Integer>(parentTestClass.makeCoeffRing(), 3, 2);
-		IPolynomialFunction<Integer> fivePowerThree = new MonomialFunction<Integer>(parentTestClass.makeCoeffRing(), 5, 3);
+		IPolynomialFunction<BigInteger> threePowerTwo = new MonomialFunction<BigInteger>(parentTestClass.makeCoeffRing(), 3, 2);
+		IPolynomialFunction<BigInteger> fivePowerThree = new MonomialFunction<BigInteger>(parentTestClass.makeCoeffRing(), 5, 3);
 
 		Assert.assertEquals(threePowerTwo, PolynomialUtil.fromString(parentTestClass.makeRing(), "3.2"));
 		Assert.assertEquals(parentTestClass.makeRing().add(fivePowerThree, threePowerTwo), PolynomialUtil.fromString(parentTestClass.makeRing(), "3.2+5.3"));
